pyproject.toml
[project]
name = "zidnahum-hudaa"
version = "1.0.0"
description = "A management system for Masjid (Muslim house of worship)."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-stubs-ext>=5.2.9",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "openpyxl>=3.1.5",
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.9",
    "djangorestframework-stubs>=3.16.8",
]

[tool.ruff]
indent-width = 2
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
